Separate experiment and animation configurations
We had been smushing the options to configure the Avida experiment
and the animation into a number of different arguments.

Now a call to plot_experiment(...) will take two required and
one optional argument: expr_cfg, which is the configuration
just for the experiment; anim_cfg, which is the configuration
just for the animation; and data_transform, which is an optional
argument to transform the data prior to sending it to be plot.

The actual function to begin plotting the animation now has
a reasonable set of defaults if the passed expr_cfg (via the
**kw argument) is missing some values.  The first positional
parameters for data and world size are required, however.