Initial commit of WebSocket playbook
The idea behind this branch is that we can play around with examining
Avida's resource system without having to be within a Python notebook.
What this will give us is the ability for novice users to play around
with settings without having to wory about the details of string
substitution, subprocesses, and function calls.

How I imagine this working: a socketio connection is established between
the server and the client page in a manner similar to that used in the
AEDBridge project.  At this point JSON packages can be sent between the
two.  The server will be responsible for serving the webpage, its
resources, and executing the python code necessary to generate the
figures as specified by the JSON packet delievered to it by its client.

This particular modification shouldn't be too much of a challenge;
getting the form data is new to me.

At the moment on the bare layout of the page is complete.