Working copy.
Blitting is implemented, and I modified the front-end of the
FactoredExperiment class to replace the child-classes that I was
envisioning.  I made this decision because it'd be prudent to see the
actual environment string itself in a notebook and just mark the
individual factors that need to be assigned values during experiment
configuration rather than hide a bunch of default settings within the
module.