Bug fixes and color blending.
Resource experiments now support the use of multiple resources.  The
process of blending the colors for each resource is terribly slow.

I also fixed a bug where the FactoredExperimentAnimation class did not
handle the case when there was only a single factor being iterated over.

Note that the name of the classes for the Single and Factored
experiments have changed.  The notebooks included in this commit have
been written to work with these new changes.