Multi-resource Fast Blending Available
With the help of Cython, color blending has improved by a factor of 100.
It is now possible to blend in under a second what it took a quarter
hour in the last version.

In order to enable this, blend.pyx has been added to this project.  It
requires a C++ compiler and the cython package be installed.  The
building process for converting blend.pyx to its native c code has been
"hidden" by use of its loader utility, so no setup is required.

I've also renamed the Multi-factor classes.  They're longer now, but
they are more descriptive.

I've updated all the notebooks to use these new classes.

I also had to add a monkey patch to get gradients to reappear in the
plots.  Seaborn had been giving me 256 levels instead of the 16 that I'd
like.  256 is too many.  Resource "waves" now show up again.