Added Masking; Preparing for Reaction Overlays
Both single and multiple-resource experiment plots now have masking
available.  This means that low-levels of the resource will now display
a light-level of their colormap color and white when no resource(s) are
available in a given cell.

I've also normalized the code between the SingleExperiment and
FactoredExperiment modules so their internal behavior and variable names
are more consistent.

All notebooks are currently working.  I removed a couple of notebooks
that are no longer necessary.

Commenting updated in FactoredExperiment module.

Because of a quirk in imshow, fully masked data cannot be plotted.  A
utility function check_masks has been added to return an all-white plot
when this is the case.

Moved our deault-heads.org to default-heads-norep.org in preparation for
when we run experiments with actual organisms.