Add export ability; make BlitArtists factories.
Both single and factored experiments now have common fields to quickly
reproduce the configuration files that were used to generate them.

Utilties now has an save_n_html function that will save a file based on
the animation's title and include metainformation about how the image
was generated (see above).

I normalized all the Python notebooks to use save_n_html to
automatically generate plots we already have and test its functionality.

BlitArtists didn't work with factored experiments as a copy of each
artist was needed for each axis.  This has now been fixed.  When a
BlitArtist is passed an axis_ndx, as in the case when it is used with
Multi-Factor experiments, a copy of itself is created and returned by
the blit_build method.  It's kind a flywheel pattern, I believe.  At any
rate, BlitArtists should work on Multi-Factored Animations now.